sdsx: some improvements
With -t the (error) messages from the client is shown directly on
console, which is especially convenient when -G is being used.
Besides, it might work better than --wait anyway.

Since we want to provide the "profile" env vars to the client with
bash -lc, let's use the command string instead of -E to provide
DISPLAY and XAUTHORITY to the client. (Although practically it
won't make any diffference in most cases, logically-speaking this
make sure the env vars set by sxrc won't be overriden the bash -l.)